modulise cluster_BGC_mash_bigscape functions
