--- pyBioinfo/pyBioinfo_modules/mash_BGC_clustering.py
@@ -1,14 +1,0 @@
-
-class mashBGC_ClusteringResult:
-    """
-    Store major results after mash clustering and medoid calculating.
-    Including:
-        clusterInfoDict: dict = pickle.load(fh)
-        family_distance_matrice: dict[str, list[list[float]]] = pickle.load(fh)
-        dict_medoids: dict[str, list[str]] = pickle.load(fh)
-
-    
-
-    """
-    def __init__(self, ) -> None:
-        pass
